refactor(protocol): split frame-bound contract tests
